chore(common): split precompiles into submodules
